Refactor parser to parser.c
